Move expression module to sexe-expression crate
