+++ Cargo.toml
@@ -4,6 +4,7 @@ version = "0.1.0"
 authors = ["Jeremy Day <jadaytime@gmail.com>"]
 
 [dependencies]
+sexe_expression = { version = "0.1.0", path = "sexe-expression" }
 nom     = "^4.1"
 termion = "^1.5"
