+++ sexe-expression/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "sexe-expression"
+name = "sexe_expression"
 version = "0.1.0"
 
 [dependencies]
