+++ src/interface.rs
@@ -53,28 +53,27 @@ impl Input for NumberInput {
                 // Reset to placeholder if our string is too short.
                 if self.display_string.len() <= 2 {
                     self.display_string = String::from("+0");
-                }
-                else {
+                } else {
                     self.display_string.pop();
                 }
-            },
+            }
             event::Key::Char(digit) if digit.is_ascii_digit() => {
                 if &self.display_string == "+0" || &self.display_string == "-0" {
                     self.display_string.pop();
                 }
                 self.display_string.push(*digit);
-            },
+            }
             event::Key::Char('+') => {
                 self.display_string.replace_range(..1, "+");
-            },
+            }
             event::Key::Char('-') => {
                 self.display_string.replace_range(..1, "-");
-            },
+            }
             event::Key::Char('.') => {
                 if !self.display_string.contains(".") {
                     self.display_string.push('.');
                 }
-            },
+            }
             _ => (),
         };
         self.number_value = self.display_string.parse().unwrap();
@@ -86,10 +85,10 @@ impl Input for TextInput {
         match key {
             event::Key::Backspace => {
                 self.string.pop();
-            },
+            }
             event::Key::Char(c) => {
                 self.string.push(*c);
-            },
+            }
             _ => (),
         };
     }
@@ -113,7 +112,12 @@ enum Error {
     RangeError,
 }
 
-fn evaluate_function_over_domain(start_x: f64, end_x: f64, resolution: u32, function_string: &str) -> Result<Vec<(f64, f64)>, Error> {
+fn evaluate_function_over_domain(
+    start_x: f64,
+    end_x: f64,
+    resolution: u32,
+    function_string: &str,
+) -> Result<Vec<(f64, f64)>, Error> {
     let mut vars_map = HashMap::new();
     vars_map.insert("x".to_string(), start_x);
 
@@ -123,23 +127,26 @@ fn evaluate_function_over_domain(start_x: f64, end_x: f64, resolution: u32, func
                 return Err(Error::ParseError);
             }
             func
-        },
+        }
         Err(_) => {
             return Err(Error::ParseError);
-        },
+        }
     };
 
     let step_width = (end_x - start_x) / resolution as f64;
 
-    Ok((0..resolution).map(|x| start_x + (x as f64 * step_width)).filter_map(|x| {
-        if let Some(val) = vars_map.get_mut(&"x".to_string()) {
-            *val = x;
-        }
-        match func.evaluate(&vars_map) {
-            Ok(y) => Some((x, y)),
-            Err(_) => None,
-        }
-    }).collect())
+    Ok((0..resolution)
+        .map(|x| start_x + (x as f64 * step_width))
+        .filter_map(|x| {
+            if let Some(val) = vars_map.get_mut(&"x".to_string()) {
+                *val = x;
+            }
+            match func.evaluate(&vars_map) {
+                Ok(y) => Some((x, y)),
+                Err(_) => None,
+            }
+        })
+        .collect())
 }
 
 impl Application {
@@ -153,13 +160,13 @@ impl Application {
                     SelectedBox::EndX => SelectedBox::StartX,
                     _ => SelectedBox::Function,
                 };
-            },
+            }
             event::Key::Right => {
                 self.selected_box = match self.selected_box {
                     SelectedBox::Function => SelectedBox::StartX,
                     _ => SelectedBox::EndX,
                 };
-            },
+            }
             // Otherwise we hand off input to the children.
             _ => match self.selected_box {
                 SelectedBox::Function => self.function_input.process_input(&key),
@@ -183,9 +190,9 @@ impl Application {
                         Paragraph::default()
                             .block(
                                 Block::default()
-                                .title("Function")
-                                .borders(Borders::ALL)
-                                .border_style(self.get_box_style(SelectedBox::Function))
+                                    .title("Function")
+                                    .borders(Borders::ALL)
+                                    .border_style(self.get_box_style(SelectedBox::Function)),
                             )
                             .style(self.get_input_style(SelectedBox::Function))
                             .wrap(false)
@@ -194,9 +201,9 @@ impl Application {
                         Paragraph::default()
                             .block(
                                 Block::default()
-                                .title("Start X")
-                                .borders(Borders::ALL)
-                                .border_style(self.get_box_style(SelectedBox::StartX))
+                                    .title("Start X")
+                                    .borders(Borders::ALL)
+                                    .border_style(self.get_box_style(SelectedBox::StartX)),
                             )
                             .style(self.get_input_style(SelectedBox::StartX))
                             .wrap(false)
@@ -205,9 +212,9 @@ impl Application {
                         Paragraph::default()
                             .block(
                                 Block::default()
-                                .title("End X")
-                                .borders(Borders::ALL)
-                                .border_style(self.get_box_style(SelectedBox::EndX))
+                                    .title("End X")
+                                    .borders(Borders::ALL)
+                                    .border_style(self.get_box_style(SelectedBox::EndX)),
                             )
                             .style(self.get_input_style(SelectedBox::EndX))
                             .wrap(false)
@@ -216,26 +223,33 @@ impl Application {
                     });
                 Chart::default()
                     .block(Block::default().title("Plot").borders(Borders::ALL))
-                    .x_axis(Axis::default()
-                        .title("X")
-                        .bounds([self.start_x_input.number_value, self.end_x_input.number_value])
-                        .labels(&[
+                    .x_axis(
+                        Axis::default()
+                            .title("X")
+                            .bounds([
+                                self.start_x_input.number_value,
+                                self.end_x_input.number_value,
+                            ])
+                            .labels(&[
                                 format!("{:.2}", self.start_x_input.number_value).as_str(),
                                 "0",
                                 format!("{:.2}", self.end_x_input.number_value).as_str(),
-                        ]))
-                    .y_axis(Axis::default()
-                        .title("Y")
-                        .bounds([self.start_y, self.end_y])
-                        .labels(&[
+                            ]),
+                    )
+                    .y_axis(
+                        Axis::default()
+                            .title("Y")
+                            .bounds([self.start_y, self.end_y])
+                            .labels(&[
                                 format!("{:.2}", self.start_y).as_str(),
                                 "0",
                                 format!("{:.2}", self.end_y).as_str(),
-                        ]))
+                            ]),
+                    )
                     .datasets(&[Dataset::default()
-                                .marker(Marker::Braille)
-                                .style(Style::default().fg(Color::Magenta))
-                                .data(&self.evaluation)])
+                        .marker(Marker::Braille)
+                        .style(Style::default().fg(Color::Magenta))
+                        .data(&self.evaluation)])
                     .render(t, &chunks[1]);
             });
 
@@ -254,10 +268,8 @@ impl Application {
         match self.plot_function() {
             Ok(vec) => {
                 // Filters all instances of f64::NAN from the vector
-                self.evaluation = vec.into_iter()
-                    .filter(|&(_,a)| a.is_normal()).collect();
-                let (start_y, end_y) =
-                    determine_y_bounds(&self.evaluation).unwrap_or((0.0, 0.0));
+                self.evaluation = vec.into_iter().filter(|&(_, a)| a.is_normal()).collect();
+                let (start_y, end_y) = determine_y_bounds(&self.evaluation).unwrap_or((0.0, 0.0));
                 if start_y == end_y {
                     let end_y_abs = end_y.abs();
                     self.start_y = -end_y_abs;
@@ -266,12 +278,12 @@ impl Application {
                     self.start_y = start_y;
                     self.end_y = end_y;
                 }
-            },
+            }
             Err(_) => {
                 self.evaluation = Vec::new();
                 self.start_y = 0.0;
                 self.end_y = 0.0;
-            },
+            }
         }
 
         self.draw(&mut terminal, &term_size);
@@ -294,8 +306,7 @@ impl Application {
             match self.plot_function() {
                 Ok(vec) => {
                     // Filters all instances of f64::NAN from the vector
-                    self.evaluation = vec.into_iter()
-                        .filter(|&(_,a)| a.is_normal()).collect();
+                    self.evaluation = vec.into_iter().filter(|&(_, a)| a.is_normal()).collect();
                     let (start_y, end_y) =
                         determine_y_bounds(&self.evaluation).unwrap_or((0.0, 0.0));
                     if start_y == end_y {
@@ -306,12 +317,12 @@ impl Application {
                         self.start_y = start_y;
                         self.end_y = end_y;
                     }
-                },
+                }
                 Err(_) => {
                     self.evaluation = Vec::new();
                     self.start_y = 0.0;
                     self.end_y = 0.0;
-                },
+                }
             }
 
             self.draw(&mut terminal, &term_size);
@@ -323,9 +334,13 @@ impl Application {
     fn plot_function(&mut self) -> Result<Vec<(f64, f64)>, Error> {
         if self.start_x_input.number_value >= self.end_x_input.number_value {
             Err(Error::RangeError)
-        }
-        else {
-            evaluate_function_over_domain(self.start_x_input.number_value, self.end_x_input.number_value, self.resolution, &self.function_input.string)
+        } else {
+            evaluate_function_over_domain(
+                self.start_x_input.number_value,
+                self.end_x_input.number_value,
+                self.resolution,
+                &self.function_input.string,
+            )
         }
     }
 
@@ -336,7 +351,7 @@ impl Application {
 
     fn get_box_style(&self, selected: SelectedBox) -> Style {
         if selected == self.selected_box {
-            Style::default().fg(Color::Magenta) 
+            Style::default().fg(Color::Magenta)
         } else {
             Style::default().fg(Color::Gray)
         }
@@ -364,4 +379,3 @@ pub fn display() {
     };
     application.start();
 }
-
